requirements.txt
google-generativeai
python-telegram-bot
python-dotenv
nest_asyncio
mcp
langchain
langchain-google-genai
